Transfer to use x86_64 crate, reimplement the frame allocator
Now we aren't really allocating on our own, just providing the needed
interface for x86_64 to work it out on its own. Really decreased the
amount of code we had, I was really getting stuck before.

As of now, we are recursively mapping the P4 table in the bootloader
script, use this in Rust code to map needed pages just by iterating over
them and checking which ones are free.

We also read the data which multiboot-compliant bootloaders provide us
with to determine which memory areas aren't taken up by the kernel or the
bootloader.